05_ACTIVE_DEV/requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
Pillow>=10.0
